+++ src/main/java/app/snapshot/qure/admin/complain/service/IComplainService.java
@@ -8,24 +8,24 @@ import app.snapshot.qure.admin.complain.dto.ComplainDto;
 
 public interface IComplainService {
 
-    List<ComplainDto> getAllComplainList(int managerId);
+    List<ComplainDto> getAllComplainList(Long managerId);
 
     List<ComplainDto> getAllComplainList(
-            int managerId, String q, String status,
+    		Long managerId, String q, String status,
             LocalDateTime fromDt, LocalDateTime toExclusive);
 
     List<ComplainDto> getAllComplainList(
-            int managerId, String q, String status,
+    		Long managerId, String q, String status,
             LocalDateTime fromDt, LocalDateTime toEx,
             int offset, int size);
 
     int countComplainList(
-            int managerId, String q, String status,
+    		Long managerId, String q, String status,
             LocalDateTime fromDt, LocalDateTime toEx);
 
-    ComplainDto getByIdForManager(int reportId, int managerId);
+    ComplainDto getByIdForManager(int reportId, Long managerId);
 
-    int deleteByIdForManager(int reportId, int managerId);
+    int deleteByIdForManager(int reportId, Long managerId);
 
-    int updateStatusForManager(int reportId, int managerId, String status);
+    int updateStatusForManager(int reportId, Long managerId, String status);
 }
